CSP-258 minor snagging from TR comments

+++ src/SFA.DAS.Roatp.Jobs/Requests/ProviderAddressLookupRequest.cs
@@ -4,6 +4,6 @@ namespace SFA.DAS.Roatp.Jobs.Requests
 {
     public class ProviderAddressLookupRequest
     {
-        public List<long> Ukprns { get; set; }
+        public List<int> Ukprns { get; set; }
     }
 }
